# Imporve perf of random read and insert compare by suggesting inlining to the compiler