@yuslepukhin 2018-03-23
This addresses https://github.com/facebook/rocksdb/issues/3590